+++ packages/@tailwindcss-upgrade/src/codemods/css/migrate-at-apply.test.ts
@@ -2,7 +2,7 @@ import { __unstable__loadDesignSystem } from '@tailwindcss/node'
 import dedent from 'dedent'
 import postcss from 'postcss'
 import { expect, it } from 'vitest'
-import type { Config } from '../../../tailwindcss/src/compat/plugin-api'
+import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
 import { migrateAtApply } from './migrate-at-apply'
 
 const css = dedent
