+++ packages/@tailwindcss-upgrade/src/codemods/css/migrate-at-layer-utilities.ts
@@ -1,8 +1,8 @@
 import { type AtRule, type Comment, type Plugin, type Rule } from 'postcss'
 import SelectorParser from 'postcss-selector-parser'
-import { segment } from '../../../tailwindcss/src/utils/segment'
-import { Stylesheet } from '../stylesheet'
-import { walk, WalkAction, walkDepth } from '../utils/walk'
+import { segment } from '../../../../tailwindcss/src/utils/segment'
+import { Stylesheet } from '../../stylesheet'
+import { walk, WalkAction, walkDepth } from '../../utils/walk'
 
 export function migrateAtLayerUtilities(stylesheet: Stylesheet): Plugin {
   function migrate(atRule: AtRule) {
