+++ packages/@tailwindcss-upgrade/src/codemods/css/migrate-media-screen.test.ts
@@ -2,7 +2,7 @@ import { __unstable__loadDesignSystem } from '@tailwindcss/node'
 import dedent from 'dedent'
 import postcss from 'postcss'
 import { expect, it } from 'vitest'
-import type { UserConfig } from '../../../tailwindcss/src/compat/config/types'
+import type { UserConfig } from '../../../../tailwindcss/src/compat/config/types'
 import { formatNodes } from './format-nodes'
 import { migrateMediaScreen } from './migrate-media-screen'
 import { sortBuckets } from './sort-buckets'
