+++ packages/@tailwindcss-upgrade/src/codemods/css/migrate-theme-to-var.test.ts
