+++ packages/@tailwindcss-upgrade/src/index.test.ts
@@ -3,8 +3,8 @@ import dedent from 'dedent'
 import path from 'node:path'
 import postcss from 'postcss'
 import { expect, it } from 'vitest'
-import { formatNodes } from './codemods/format-nodes'
-import { sortBuckets } from './codemods/sort-buckets'
+import { formatNodes } from './codemods/css/format-nodes'
+import { sortBuckets } from './codemods/css/sort-buckets'
 import { migrateContents } from './migrate'
 
 const css = dedent
