+++ packages/@tailwindcss-upgrade/src/index.ts
@@ -4,8 +4,10 @@ import { globby } from 'globby'
 import fs from 'node:fs/promises'
 import path from 'node:path'
 import postcss from 'postcss'
-import { formatNodes } from './codemods/format-nodes'
-import { sortBuckets } from './codemods/sort-buckets'
+import { formatNodes } from './codemods/css/format-nodes'
+import { sortBuckets } from './codemods/css/sort-buckets'
+import { migrate as migrateTemplate } from './codemods/template/migrate'
+import { prepareConfig } from './codemods/template/prepare-config'
 import { help } from './commands/help'
 import {
   analyze as analyzeStylesheets,
@@ -17,8 +19,6 @@ import { migrateJsConfig } from './migrate-js-config'
 import { migratePostCSSConfig } from './migrate-postcss'
 import { migratePrettierPlugin } from './migrate-prettier'
 import { Stylesheet } from './stylesheet'
-import { migrate as migrateTemplate } from './template/migrate'
-import { prepareConfig } from './template/prepare-config'
 import { args, type Arg } from './utils/args'
 import { isRepoDirty } from './utils/git'
 import { hoistStaticGlobParts } from './utils/hoist-static-glob-parts'
