+++ packages/@tailwindcss-upgrade/src/migrate.ts
@@ -6,19 +6,19 @@ import type { Config } from '../../tailwindcss/src/compat/plugin-api'
 import type { DesignSystem } from '../../tailwindcss/src/design-system'
 import { DefaultMap } from '../../tailwindcss/src/utils/default-map'
 import { segment } from '../../tailwindcss/src/utils/segment'
-import { migrateAtApply } from './codemods/migrate-at-apply'
-import { migrateAtLayerUtilities } from './codemods/migrate-at-layer-utilities'
-import { migrateConfig } from './codemods/migrate-config'
-import { migrateImport } from './codemods/migrate-import'
-import { migrateMediaScreen } from './codemods/migrate-media-screen'
-import { migrateMissingLayers } from './codemods/migrate-missing-layers'
-import { migratePreflight } from './codemods/migrate-preflight'
-import { migrateTailwindDirectives } from './codemods/migrate-tailwind-directives'
-import { migrateThemeToVar } from './codemods/migrate-theme-to-var'
-import { migrateVariantsDirective } from './codemods/migrate-variants-directive'
+import { migrateAtApply } from './codemods/css/migrate-at-apply'
+import { migrateAtLayerUtilities } from './codemods/css/migrate-at-layer-utilities'
+import { migrateConfig } from './codemods/css/migrate-config'
+import { migrateImport } from './codemods/css/migrate-import'
+import { migrateMediaScreen } from './codemods/css/migrate-media-screen'
+import { migrateMissingLayers } from './codemods/css/migrate-missing-layers'
+import { migratePreflight } from './codemods/css/migrate-preflight'
+import { migrateTailwindDirectives } from './codemods/css/migrate-tailwind-directives'
+import { migrateThemeToVar } from './codemods/css/migrate-theme-to-var'
+import { migrateVariantsDirective } from './codemods/css/migrate-variants-directive'
+import { detectConfigPath } from './codemods/template/prepare-config'
 import type { JSConfigMigration } from './migrate-js-config'
 import { Stylesheet, type StylesheetConnection, type StylesheetId } from './stylesheet'
-import { detectConfigPath } from './template/prepare-config'
 import { error, highlight, relative, success } from './utils/renderer'
 import { resolveCssId } from './utils/resolve'
 import { walk, WalkAction } from './utils/walk'
